+++ plugins/params/link.php
@@ -44,6 +44,29 @@ function wolf_core_link_params() {
 					'label'       => esc_html__( 'URL', 'wolf-core' ),
 					'param_name'  => 'link',
 				),
+
+				array(
+					'type'         => 'colorpicker',
+					'label'        => esc_html__( 'Text Color', 'wolf-core' ),
+					'param_name'   => 'blockquote_text_color',
+					'selectors'    => array(
+						'{{WRAPPER}} .wolf-core-link' => 'color: {{VALUE}}!important;',
+						//'{{WRAPPER}} .wolf-core-link blockquote' => 'color: {{VALUE}}!important;',
+					),
+					'page_builder' => 'elementor',
+					'group'        => esc_html__( 'Style', 'wolf-core' ),
+				),
+
+				array(
+					'type'         => 'colorpicker',
+					'label'        => esc_html__( 'Background Color', 'wolf-core' ),
+					'param_name'   => 'custom_background_color', // backward compatiblity name (WVC plugin).
+					'page_builder' => 'elementor',
+					'selectors'    => array(
+						'{{WRAPPER}} .wolf-core-link' => 'background-color: {{VALUE}}!important;',
+					),
+					'group'        => esc_html__( 'Style', 'wolf-core' ),
+				),
 			),
 		)
 	);